Add files via upload
Aquel Lanka is a small business that sells motorcycle spare parts. The shop is located in Rathmalana. I have developed a customized desktop inventory management software using C# .net. The application is capable of doing day-to-day activities in the shop including inventory management, billing, and reporting. The system used SQL Server for database management and SAP Crystal reports for reporting.